refactor: put services into folders
