+++ src/app/interfaces/user.model.ts
@@ -1,4 +1,4 @@
-export interface User {
+export interface UserDto {
   name: string;
   login: string;
   email: string;
