+++ src/app/services/user/user.service.ts
@@ -1,16 +1,15 @@
-import {Injectable, inject} from '@angular/core';
-import {HttpClient} from '@angular/common/http';
-import {User} from '../interfaces/user.model';
+import { Injectable, inject } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { UserDto } from '../../interfaces/user.model';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
-
 export class UserService {
   private apiUrl = 'https://be-12092025.onrender.com';
   private http = inject(HttpClient);
 
-  postUser(obj: User) {
+  postUser(obj: UserDto) {
     console.log(obj);
     return this.http.post(`${this.apiUrl}/auth/signup`, obj);
   }
